+++ test/com/onsoft/sandcat/jcad/decorators/InitDecoratorTest.ts
@@ -56,13 +56,13 @@ export class InitDecoratorTest {
   }
 
   @Test({
-    description: "should return the reference to the target instance"
+    description: "should return the reference to the PropertyDescriptor instance"
   })
   public decorateTargetTest():void {
     let target:any = this.decorator.decorate(
       utils.TARGET, utils.KEY, utils.DESCRIPTOR
     );
-    expect(target).to.equal(utils.TARGET);
+    expect(target).to.equal(utils.DESCRIPTOR);
   }
   
   @Test({